requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
psycopg2-binary
python-multipart
bcrypt==3.2.0
aiofiles
pytest
httpx
selenium